+++ __tests__/demo/demo.js
@@ -39,7 +39,9 @@ import {
   DetailPanelIssuesProgrammaticallyHidingWhenOpen,
   EventTargetErrorOnRowClick,
   SelectionOnRowClick,
-  DetailPanelRemounting
+  DetailPanelRemounting,
+  TreeData,
+  TableWithSummary
 } from './demo-components';
 import { I1353, I1941, I122 } from './demo-components/RemoteData';
 
@@ -71,14 +73,14 @@ render(
     <BasicRef />
 
     {/*
-    <h1>Export Data</h1>
-    <ExportData />
-    */}
+     <h1>Export Data</h1>
+     <ExportData />
+     */}
 
     {/*
-    <h1>Custom Export </h1>
-    <CustomExport />
-    */}
+     <h1>Custom Export </h1>
+     <CustomExport />
+     */}
 
     <h1>Bulk Edit</h1>
     <BulkEdit />
@@ -121,7 +123,10 @@ render(
 
     <h1>Persistent Groupings unshared</h1>
     <PersistentGroupings persistentGroupingsId="some-other-id" />
-
+    <h1>Tree data</h1>
+    <TreeData />
+    <h1>Table with Summary Row</h1>
+    <TableWithSummary />
     <h1>Remote Data Related</h1>
     <ol>
       <li>