+++ src/material-table.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { debounce } from 'debounce';
 import equal from 'fast-deep-equal/react';
-import cloneDeep from 'lodash/cloneDeep';
 import {
   Table,
   TableFooter,
@@ -96,37 +95,29 @@ export default class MaterialTable extends React.Component {
           : '';
     }
 
-    const columnsCopy = cloneDeep(props.columns);
-
+    const savedColumns = {};
     if (props.options.persistentGroupingsId) {
       let materialTableGroupings = localStorage.getItem(
         'material-table-groupings'
       );
-
       if (materialTableGroupings) {
         materialTableGroupings = JSON.parse(materialTableGroupings);
 
         if (materialTableGroupings[props.options.persistentGroupingsId]) {
           materialTableGroupings[props.options.persistentGroupingsId].forEach(
             (savedGrouping) => {
-              const column = columnsCopy.find(
-                (col) => col.field === savedGrouping.field
-              );
-              if (column) {
-                if (!column.tableData) {
-                  column.tableData = {};
-                }
-                column.tableData.groupOrder = savedGrouping.groupOrder;
-                column.tableData.groupSort = savedGrouping.groupSort;
-                column.tableData.columnOrder = savedGrouping.columnOrder;
-              }
+              savedColumns[savedGrouping.field] = {
+                groupOrder: savedGrouping.groupOrder,
+                groupSort: savedGrouping.groupSort,
+                columnOrder: savedGrouping.columnOrder
+              };
             }
           );
         }
       }
     }
 
-    this.dataManager.setColumns(columnsCopy, prevColumns);
+    this.dataManager.setColumns(props.columns, prevColumns, savedColumns);
     this.dataManager.setDefaultExpanded(props.options.defaultExpanded);
     this.dataManager.changeRowEditing();
 
@@ -199,7 +190,9 @@ export default class MaterialTable extends React.Component {
         process.env.NODE_ENV === 'development' &&
         columnPropsChanged &&
         !this.checkedForFunctions &&
-        prevProps.columns.length !== 0
+        prevProps.columns.length !== 0 &&
+        props.data[0] &&
+        props.data[0].id !== undefined
       ) {
         const bothContainFunctions =
           fixedPropsColumns.some((column) =>
@@ -313,6 +306,10 @@ export default class MaterialTable extends React.Component {
           disabled: !!this.dataManager.lastEditingRow,
           onClick: () => {
             this.dataManager.changeRowEditing();
+            if (this.state.showAddRow) {
+              this.props.editable.onRowAddCancelled &&
+                this.props.editable.onRowAddCancelled();
+            }
             this.setState({
               ...this.dataManager.getRenderState(),
               showAddRow: !this.state.showAddRow
@@ -753,7 +750,9 @@ export default class MaterialTable extends React.Component {
       query.page = 0;
       query.search = searchText;
 
-      this.onQueryChange(query);
+      this.onQueryChange(query, () => {
+        this.props.onSearchChange && this.props.onSearchChange(searchText);
+      });
     } else {
       this.setState(this.dataManager.getRenderState(), () => {
         this.props.onSearchChange && this.props.onSearchChange(searchText);
@@ -778,7 +777,9 @@ export default class MaterialTable extends React.Component {
           value: a.tableData.filterValue
         }));
 
-      this.onQueryChange(query);
+      this.onQueryChange(query, () => {
+        this.props.onFilterChange && this.props.onFilterChange(query.filters);
+      });
     } else {
       this.setState(this.dataManager.getRenderState(), () => {
         if (this.props.onFilterChange) {