+++ src/utils/data-manager.js
@@ -78,16 +78,29 @@ export default class DataManager {
       if (tableData.checked) {
         this.selectedCount++;
       }
-      return {
+      const newRow = {
         ...row,
         tableData
       };
+      if (
+        this.lastDetailPanelRow &&
+        this.lastDetailPanelRow.tableData === prevTableData
+      ) {
+        this.lastDetailPanelRow = newRow;
+      }
+      if (
+        this.lastEditingRow &&
+        this.lastEditingRow.tableData === prevTableData
+      ) {
+        this.lastEditingRow = newRow;
+      }
+      return newRow;
     });
 
     this.filtered = false;
   }
 
-  setColumns(columns, prevColumns = []) {
+  setColumns(columns, prevColumns = [], savedColumns = {}) {
     let usedWidth = ['0px'];
 
     this.columns = columns.map((columnDef, index) => {
@@ -104,6 +117,7 @@ export default class DataManager {
         usedWidth.push(width);
       }
       const prevColumn = prevColumns.find(({ id }) => id === index);
+      const savedColumnTableData = savedColumns[columnDef.field] ?? {};
       const tableData = {
         columnOrder: index,
         filterValue: columnDef.defaultFilter,
@@ -112,6 +126,7 @@ export default class DataManager {
         width,
         initialWidth: width,
         additionalWidth: 0,
+        ...savedColumnTableData,
         ...(prevColumn ? prevColumn.tableData : {}),
         ...columnDef.tableData,
         id: index
@@ -538,12 +553,13 @@ export default class DataManager {
             result &&
             result.tableData &&
             result.tableData.childRows &&
-            result.tableData.childRows[current]
+            result.tableData.childRows.find(
+              (row) => row && row.tableData.uuid === current
+            )
           );
         },
         { tableData: { childRows: renderData } }
       );
-
       return node;
     } else {
       const data = { groups: renderData };
@@ -568,7 +584,6 @@ export default class DataManager {
       if (!result) {
         return undefined;
       }
-
       if (result.groupsIndex[current] !== undefined) {
         return result.groups[result.groupsIndex[current]];
       }
@@ -918,7 +933,6 @@ export default class DataManager {
     this.treefiedData = [];
     this.treefiedDataLength = 0;
     this.treeDataMaxLevel = 0;
-
     // if filter or search is enabled, collapse the tree
     if (
       this.searchText ||
@@ -943,10 +957,10 @@ export default class DataManager {
         }
 
         addRow(parent);
-
         rowData.tableData.path = [
           ...parent.tableData.path,
-          parent.tableData.childRows.length - 1
+          parent.tableData.childRows[parent.tableData.childRows.length - 1]
+            .tableData.uuid
         ];
         this.treeDataMaxLevel = Math.max(
           this.treeDataMaxLevel,
@@ -956,7 +970,7 @@ export default class DataManager {
         if (!this.treefiedData.includes(rowData)) {
           this.treefiedData.push(rowData);
           this.treefiedDataLength++;
-          rowData.tableData.path = [this.treefiedData.length - 1];
+          rowData.tableData.path = [rowData.tableData.uuid];
         }
       }
     };